IAM/iam_service/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
motor>=2.5.1
pymongo>=3.12.0
pydantic>=1.8.2
pydantic-settings>=2.9.1
python-jose>=3.3.0
passlib>=1.7.4
python-dotenv>=0.19.0
email-validator>=1.1.3
python-multipart>=0.0.5
bcrypt>=3.2.0